pyproject.toml
[tool.poetry]
name = "gan"
version = "0.1.0"
description = ""
authors = ["Mikedev <mdv1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.0"
pytorch-lightning = "^1.7.0"
path = "^16.4.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

